% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myverse.R
\docType{package}
\name{myverse-package}
\alias{myverse}
\alias{myverse-package}
\title{myverse: Easily Load a Set of Packages}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Define a list of packages as options and attach them easily by attaching
    the 'myverse' package. This is arguably better than using the RProfile because
    you need to opt in in each session so this won't damage reproductibility, moreover
    this prints the list of attached packages and their versions. This is directly
    inspired by the Hadley Wickham's 'tidyverse' package and borrows some of its
    code and look.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/moodymudskipper/myverse}
  \item Report bugs at \url{https://github.com/moodymudskipper/myverse/issues}
}

}
\author{
\strong{Maintainer}: Antoine Fabri \email{antoine.fabri@gmail.com}

Other contributors:
\itemize{
  \item Hadley Wickham \email{hadley@rstudio.com} (Author of the tidyverse package which this package's code derives from) [contributor]
  \item RStudio (Copyright holder of the tidyverse package) [contributor]
}

}
\keyword{internal}
