% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{myverse_conflicts}
\alias{myverse_conflicts}
\title{Conflicts between the myverse packages and other packages}
\usage{
myverse_conflicts(cluster = NULL)
}
\arguments{
\item{cluster}{Name of the cluster to consider, by default if `getOption("myverse.pkgs")`
is a vector it's taken as is, and if it's a list, the first element
(presumably a vector) is taken.}
}
\description{
This function lists all the conflicts between packages in the myverse
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from dplyr. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
myverse_conflicts()
}
