% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{myverse_packages}
\alias{myverse_packages}
\title{List all packages in the myverse}
\usage{
myverse_packages(include_self = TRUE, cluster = NULL)
}
\arguments{
\item{include_self}{Include tidyverse in the list?}

\item{cluster}{Name of the cluster to consider, by default if `getOption("myverse.pkgs")`
is a vector it's taken as is, and if it's a list, the first element
(presumably a vector) is taken.}
}
\description{
List all packages in the myverse
}
\examples{
myverse_packages()
}
