% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{myverse_sitrep}
\alias{myverse_sitrep}
\title{Get a situation report on the myverse}
\usage{
myverse_sitrep(cluster = NULL)
}
\arguments{
\item{cluster}{Name of the cluster to consider, by default if `getOption("myverse.pkgs")`
is a vector it's taken as is, and if it's a list, the first element
(presumably a vector) is taken.}
}
\description{
This function gives a quick overview of the versions of R and RStudio as
well as all myverse packages. It's primarily designed to help you get
a quick idea of what's going on when you're helping someone else debug
a problem.
}
