% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{myverse_update}
\alias{myverse_update}
\title{Update myverse packages}
\usage{
myverse_update(recursive = FALSE, repos = getOption("repos"), cluster = NULL)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
tidyverse packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}

\item{cluster}{Name of the cluster to consider, by default if `getOption("myverse.pkgs")`
is a vector it's taken as is, and if it's a list, the first element
(presumably a vector) is taken.}
}
\description{
This will check to see if all myverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
tidyverse_update()
}
}
